PlantMonitor/pyproject.toml
[project]
name = "groupwork"
version = "0.1.0"
description = "DynamoDBデータの可視化アプリケーション"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.37.33",
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "python-dotenv>=1.1.0",
    "uvicorn[standard]>=0.34.0",
]
